Update fizzbuzz to match test data
COBOL is strict about the width of a number based on the digits in the
PIC clause. In this case PIC 99 displays a two-digit number i.e. "16"
normally, but a single-digit number i.e. 7 would display as " 7" in the
output. So I had to create a PIC clause for each width, because TRIM
didn't work as I was expecting. I probably called it incorrectly.